Refactor SSR into separate helper function
